requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
paddlepaddle>=2.5.0
paddlepaddle-gpu>=2.5.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Metrics and utilities
torchmetrics>=1.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0